feat: Was trying to figure out why authentication was being persisted in browser between reloads
I didn't add in any code for cookies, but upon refreshing (and not having signed out yet), the sign out button, indicating I was still signed in, reappeared. This, I think, is because Google's sign in button makes a request right away if it's on the page, in such a way that if the user hasn't signed out yet, the user's sign in credentials are returned. Because of this, this.props.auth changed due to the on-datasuccess property on the google butotn, which forced a render() invocation, showing the sign out button due to the props state change.